chore: move tests to slim down sloc
