+++ contract/vault/src/impl_callbacks.rs
@@ -671,13 +671,13 @@ impl Contract {
     }
 }
 
-pub(crate) struct SupplyReconciliation {
-    new_principal: u128,
-    accepted_event: u128,
-    remaining: u128,
+pub struct SupplyReconciliation {
+    pub new_principal: u128,
+    pub accepted_event: u128,
+    pub remaining: u128,
 }
 
-pub(crate) fn reconcile_supply_outcome(
+pub fn reconcile_supply_outcome(
     total_position: &u128,
     before: &u128,
     remaining: &u128,
@@ -699,7 +699,7 @@ pub struct WithdrawReconciliation {
 }
 
 /// Pure reconciliation for withdraw read outcome to enable unit tests
-pub(crate) fn reconcile_withdraw_outcome(
+pub fn reconcile_withdraw_outcome(
     before_principal: u128,
     new_principal: u128,
     remaining_total: u128,
@@ -715,900 +715,5 @@ pub(crate) fn reconcile_withdraw_outcome(
         remaining_next,
         collected_next,
         idle_delta,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::u128;
-
-    use crate::impl_callbacks::reconcile_supply_outcome;
-    use crate::test_utils::*;
-
-    use near_sdk::json_types::U128;
-    use near_sdk::test_utils::accounts;
-    use near_sdk::PromiseOrValue;
-    use near_sdk::PromiseResult;
-    use near_sdk_contract_tools::ft::Nep141 as _;
-    use rstest::rstest;
-
-    use crate::Contract;
-    use near_sdk::AccountId;
-    use rstest::fixture;
-    use templar_common::vault::Error;
-    use templar_common::vault::OpState;
-
-    #[fixture]
-    fn vault_id() -> AccountId {
-        accounts(0)
-    }
-
-    #[fixture]
-    fn c(vault_id: AccountId) -> Contract {
-        setup_env(&vault_id, &vault_id, vec![]);
-        new_test_contract(&vault_id)
-    }
-
-    // Contract with the env used by after_supply_1_check_* tests
-    #[fixture]
-    fn c_max(vault_id: AccountId) -> Contract {
-        setup_env(
-            &vault_id,
-            &vault_id,
-            vec![PromiseResult::Successful(
-                near_sdk::serde_json::to_vec(&U128(u128::MAX))
-                    .unwrap_or_else(|e| near_sdk::env::panic_str(&e.to_string())),
-            )],
-        );
-        new_test_contract(&vault_id)
-    }
-
-    #[fixture]
-    fn receiver() -> AccountId {
-        mk(9)
-    }
-
-    #[fixture]
-    fn owner() -> AccountId {
-        accounts(1)
-    }
-
-    #[rstest]
-    fn after_supply_1_check_allocating_not_allocating(mut c_max: Contract) {
-        let mut c = c_max;
-
-        c.op_state = OpState::Idle;
-
-        c.after_supply_1_check(Ok(U128(1)), 0, 2, Default::default());
-
-        assert_eq!(c.op_state, OpState::Idle);
-        assert_eq!(c.plan, None);
-    }
-
-    #[test]
-    fn after_supply_1_check_allocating_not_allocating_index() {
-        let vault_id = accounts(0);
-        setup_env(
-            &vault_id,
-            &vault_id,
-            vec![PromiseResult::Successful(
-                near_sdk::serde_json::to_vec(&U128(u128::MAX))
-                    .unwrap_or_else(|e| near_sdk::env::panic_str(&e.to_string())),
-            )],
-        );
-
-        let mut c = new_test_contract(&vault_id);
-
-        let op_id = 1;
-        let receiver = mk(7);
-
-        c.op_state = OpState::Allocating {
-            op_id,
-            index: 0u32,
-            remaining: 0u128,
-        };
-
-        c.after_supply_1_check(Ok(U128(1)), op_id + 1, 0, Default::default());
-
-        assert_eq!(c.op_state, OpState::Idle);
-        assert_eq!(c.plan, None);
-    }
-
-    #[test]
-    fn after_supply_1_check_allocating() {
-        let vault_id = accounts(0);
-        setup_env(
-            &vault_id,
-            &vault_id,
-            vec![PromiseResult::Successful(
-                near_sdk::serde_json::to_vec(&U128(u128::MAX))
-                    .unwrap_or_else(|e| near_sdk::env::panic_str(&e.to_string())),
-            )],
-        );
-
-        let mut c = new_test_contract(&vault_id);
-
-        let op_id = 1;
-        let receiver = mk(7);
-
-        c.op_state = OpState::Allocating {
-            op_id,
-            index: 0u32,
-            remaining: 0u128,
-        };
-
-        c.after_supply_1_check(Ok(U128(1)), op_id, 0, Default::default());
-
-        assert_eq!(c.op_state, OpState::Idle);
-        assert_eq!(c.plan, None);
-    }
-
-    #[test]
-    fn after_send_to_user_success_no_escrow() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-
-        let mut c = new_test_contract(&vault_id);
-
-        let receiver = mk(7);
-
-        c.idle_balance = 1_000;
-        c.op_state = OpState::Payout {
-            op_id: 1,
-            receiver: receiver.clone(),
-            amount: 200,
-            owner: accounts(1),
-            escrow_shares: 0,
-            burn_shares: 0,
-        };
-
-        let ok = c.after_send_to_user(Ok(()), 1, receiver.clone(), U128(200));
-        assert!(ok, "Payout should report success");
-        assert_eq!(c.idle_balance, 800, "Idle balance must decrease by payout");
-        assert!(
-            matches!(c.op_state, OpState::Idle),
-            "Vault must go Idle after successful payout"
-        );
-    }
-
-    #[rstest]
-    fn after_exec_withdraw_read_none_to_payout(mut c: Contract) {
-        // Prepare a single-market withdraw queue with non-zero principal
-        let market = mk(8);
-        c.withdraw_queue.push(market.clone());
-        c.market_supply.insert(market.clone(), 100);
-
-        // Withdrawing: need 60, already collected 10; expect position None => new_principal = 0, withdrawn = 100, credited = min(100, 60) = 60
-        c.op_state = OpState::Withdrawing {
-            op_id: 42,
-            index: 0,
-            remaining: 60,
-            receiver: mk(9),
-            collected: 10,
-            owner: accounts(1),
-            escrow_shares: 50,
-        };
-
-        let res = c.after_exec_withdraw_read(Ok(None), 42, 0, U128(100), U128(60));
-
-        match res {
-            PromiseOrValue::Promise(p) => {}
-            _ => panic!("Expected a Promise to send payout"),
-        }
-
-        assert_eq!(
-            *c.market_supply.get(&market).unwrap_or(&u128::MAX),
-            0,
-            "Market principal should be updated to 0"
-        );
-
-        assert_eq!(
-            c.idle_balance, 100,
-            "Idle balance should increase by returned amount"
-        );
-
-        // State should transition to Payout with amount = collected (10) + credited (60) = 70
-        match &c.op_state {
-            OpState::Payout { amount, .. } => {
-                assert_eq!(*amount, 70, "Payout amount must match collected + credited");
-            }
-            other => panic!("Unexpected state after read: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn after_skim_balance_zero_noop() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-
-        let mut c = new_test_contract(&vault_id);
-
-        let res = c.after_skim_balance(Ok(U128(0)), mk(10), mk(11));
-        match res {
-            PromiseOrValue::Value(()) => {}
-            _ => panic!("Skim with zero balance must be a no-op"),
-        }
-    }
-
-    #[test]
-    fn after_skim_balance_positive_returns_promise() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-
-        let mut c = new_test_contract(&vault_id);
-
-        // Positive balance -> Promise to ft_transfer
-        let res = c.after_skim_balance(Ok(U128(123)), mk(10), mk(11));
-        match res {
-            PromiseOrValue::Promise(_) => { //NOTE: one day we will be able to read the promise
-                 //definition :<
-            }
-            _ => panic!("Skim with positive balance must return a Promise"),
-        }
-    }
-
-    /// Property: Payout failure keeps idle_balance unchanged and does not burn escrow
-    #[rstest(
-        idle => [0u128, 1, 100],
-        escrow => [0u128, 1, 50],
-        amount => [0u128, 1, 25]
-    )]
-    fn prop_after_send_to_user_failure_keeps_idle(idle: u128, escrow: u128, amount: u128) {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        let receiver = mk(7);
-        let owner = accounts(1);
-
-        if escrow > 0 {
-            use near_sdk_contract_tools::ft::Nep141Controller as _;
-
-            c.deposit_unchecked(&near_sdk::env::current_account_id(), escrow)
-                .unwrap_or_else(|e| near_sdk::env::panic_str(&e.to_string()));
-        }
-
-        c.idle_balance = idle;
-        c.op_state = OpState::Payout {
-            op_id: 1,
-            receiver: receiver.clone(),
-            amount,
-            owner: owner.clone(),
-            escrow_shares: escrow,
-            burn_shares: escrow,
-        };
-
-        let before = c.idle_balance;
-        let ok = c.after_send_to_user(
-            Err(near_sdk::PromiseError::Failed),
-            1,
-            receiver.clone(),
-            U128(amount),
-        );
-        assert!(!ok, "Payout failure should return false");
-        assert_eq!(
-            c.idle_balance, before,
-            "idle_balance must stay the same on payout failure"
-        );
-        assert!(
-            matches!(c.op_state, OpState::Idle),
-            "Vault must go Idle after payout failure"
-        );
-    }
-
-    /// Property: Create-withdraw failure skips to next market and if collected>0 ends in Payout
-    #[rstest(
-        collected => [1u128, 10u128],
-        need => [1u128, 5u128]
-    )]
-    fn prop_after_create_withdraw_req_failure_skips(collected: u128, need: u128) {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        // Single-market queue so advancing index reaches end-of-queue
-        let market = mk(8);
-        c.withdraw_queue.push(market.clone());
-        c.market_supply.insert(market.clone(), 100);
-
-        c.op_state = OpState::Withdrawing {
-            op_id: 7,
-            index: 0,
-            remaining: need,
-            receiver: mk(9),
-            collected,
-            owner: accounts(1),
-            escrow_shares: 0,
-        };
-
-        let res =
-            c.after_create_withdraw_req(Err(near_sdk::PromiseError::Failed), 7, 0, U128(need));
-        match res {
-            PromiseOrValue::Promise(_) => {}
-            _ => panic!("Expected Promise after skipping to payout at end-of-queue"),
-        }
-
-        match &c.op_state {
-            OpState::Payout { amount, .. } => {
-                assert_eq!(*amount, collected, "Payout amount must equal collected");
-            }
-            other => panic!("Unexpected state: {other:?}"),
-        }
-    }
-
-    /// Property: Exec-withdraw read failure assumes unchanged principal and does not credit idle
-    #[rstest(
-        before => [0u128, 1u128, 100u128],
-        need => [0u128, 1u128, 50u128],
-        collected => [1u128, 2u128]
-    )]
-    fn prop_after_exec_withdraw_read_err_no_change(before: u128, need: u128, collected: u128) {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        let market = mk(8);
-        c.withdraw_queue.push(market.clone());
-        c.market_supply.insert(market.clone(), before);
-
-        let initial_idle = c.idle_balance;
-
-        c.op_state = OpState::Withdrawing {
-            op_id: 99,
-            index: 0,
-            remaining: need,
-            receiver: mk(9),
-            collected,
-            owner: accounts(1),
-            escrow_shares: 0,
-        };
-
-        let res = c.after_exec_withdraw_read(
-            Err(near_sdk::PromiseError::Failed),
-            99,
-            0,
-            U128(before),
-            U128(need),
-        );
-        match res {
-            PromiseOrValue::Promise(_) => {}
-            _ => panic!("Expected Promise to send payout at end-of-queue"),
-        }
-
-        assert_eq!(
-            *c.market_supply.get(&market).unwrap_or(&u128::MAX),
-            before,
-            "principal must remain unchanged on read failure"
-        );
-        assert_eq!(
-            c.idle_balance, initial_idle,
-            "idle_balance must not change when nothing credited"
-        );
-
-        match &c.op_state {
-            OpState::Payout { amount, .. } => {
-                assert_eq!(*amount, collected, "Payout amount must equal collected");
-            }
-            other => panic!("Unexpected state: {other:?}"),
-        }
-    }
-
-    /// Property: Callbacks must match current op_id or index; otherwise stop and go Idle
-    #[rstest(
-        pass_op => [false, true],
-        pass_index => [false, true]
-    )]
-    fn prop_after_exec_withdraw_read_requires_current_state(pass_op: bool, pass_index: bool) {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        let market = mk(8);
-        c.withdraw_queue.push(market.clone());
-        c.market_supply.insert(market.clone(), 10);
-
-        let real_op = 5u64;
-        let real_idx = 0u32;
-
-        c.op_state = OpState::Withdrawing {
-            op_id: real_op,
-            index: real_idx,
-            remaining: 1,
-            receiver: mk(9),
-            collected: 1,
-            owner: accounts(1),
-            escrow_shares: 0,
-        };
-
-        let call_op = if pass_op { real_op } else { real_op + 1 };
-        let call_idx = if pass_index { real_idx } else { real_idx + 1 };
-
-        let r = c.after_exec_withdraw_read(Ok(None), call_op, call_idx, U128(10), U128(1));
-        if let (true, true) = (pass_op, pass_index) {
-            assert!(
-                !matches!(c.op_state, OpState::Idle),
-                "Valid callback should not immediately stop"
-            );
-        } else {
-            // Any mismatch should stop and go Idle
-            if let PromiseOrValue::Value(()) = r {}
-            assert!(
-                matches!(c.op_state, OpState::Idle),
-                "Mismatched callback must stop and go Idle"
-            );
-        }
-    }
-
-    #[test]
-    fn refund_path_consistency() {
-        use near_sdk_contract_tools::ft::Nep141Controller as _;
-
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        // Seed escrowed shares into the vault's own account
-        let owner = accounts(1);
-        c.deposit_unchecked(&near_sdk::env::current_account_id(), 10)
-            .unwrap_or_else(|e| near_sdk::env::panic_str(&e.to_string()));
-
-        // Single-market withdraw queue (not used functionally here, just to satisfy path)
-        let market = mk(12);
-        c.withdraw_queue.push(market);
-
-        // Withdrawing state with remaining=0 and collected=0 forces refund path
-        c.op_state = OpState::Withdrawing {
-            op_id: 77,
-            index: 0,
-            remaining: 0,
-            receiver: mk(9),
-            collected: 0,
-            owner: owner.clone(),
-            escrow_shares: 10,
-        };
-
-        let supply_before = c.total_supply();
-        let vault_before = c.balance_of(&near_sdk::env::current_account_id());
-        let owner_before = c.balance_of(&owner);
-
-        // Read result with need=0 ensures credited=0; triggers refund branch
-        let res = c.after_exec_withdraw_read(Ok(None), 77, 0, U128(0), U128(0));
-        match res {
-            PromiseOrValue::Value(()) => {}
-            _ => panic!("Expected Value(()) on immediate escrow refund"),
-        }
-
-        // No burn/mint => total supply unchanged
-        assert_eq!(
-            c.total_supply(),
-            supply_before,
-            "no supply change on refund"
-        );
-        // Escrow shares transferred back to owner
-        assert_eq!(
-            c.balance_of(&near_sdk::env::current_account_id()),
-            vault_before.saturating_sub(10),
-            "vault should lose refunded escrow"
-        );
-        assert_eq!(
-            c.balance_of(&owner),
-            owner_before.saturating_add(10),
-            "owner should receive refunded escrow"
-        );
-        // Vault returns to Idle
-        assert!(
-            matches!(c.op_state, OpState::Idle),
-            "Vault must go Idle after refund"
-        );
-    }
-
-    #[test]
-    fn ctx_allocating_ok_and_err() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        c.op_state = OpState::Allocating {
-            op_id: 42,
-            index: 3,
-            remaining: 77,
-        };
-
-        let ok = c.ctx_allocating(42).expect("ctx_allocating should succeed");
-        assert_eq!(ok, (3, 77));
-
-        // Wrong op_id => error
-        assert!(c.ctx_allocating(43).is_err());
-    }
-
-    #[test]
-    fn ctx_withdrawing_ok_and_err() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        let recv = mk(1);
-        let owner = accounts(1);
-
-        c.op_state = OpState::Withdrawing {
-            op_id: 7,
-            index: 1,
-            remaining: 50,
-            receiver: recv.clone(),
-            collected: 5,
-            owner: owner.clone(),
-            escrow_shares: 10,
-        };
-
-        let (idx, rem, r, coll, o, escrow) = c
-            .ctx_withdrawing(7)
-            .expect("ctx_withdrawing should succeed");
-        assert_eq!(idx, 1);
-        assert_eq!(rem, 50);
-        assert_eq!(r, recv);
-        assert_eq!(coll, 5);
-        assert_eq!(o, owner);
-        assert_eq!(escrow, 10);
-
-        // Wrong op_id => error
-        assert!(c.ctx_withdrawing(8).is_err());
-    }
-
-    #[test]
-    fn resolve_market_helpers_supply_and_withdraw() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        // Prepare markets
-        let m1 = mk(1001);
-        let m2 = mk(1002);
-
-        // Supply: plan takes precedence
-        c.plan = Some(vec![(m2.clone(), 1u128)]);
-        c.supply_queue.push(m1.clone());
-        c.supply_queue.push(m2.clone());
-
-        assert_eq!(c.resolve_supply_market(0).unwrap(), m2);
-        assert!(matches!(
-            c.resolve_supply_market(1),
-            Err(Error::MissingMarket(1))
-        ));
-
-        // Without plan, use queue
-        c.plan = None;
-        assert_eq!(c.resolve_supply_market(0).unwrap(), m1);
-        assert_eq!(c.resolve_supply_market(1).unwrap(), m2);
-        assert!(matches!(
-            c.resolve_supply_market(2),
-            Err(Error::MissingMarket(2))
-        ));
-
-        // Withdraw resolver uses withdraw_queue
-        c.withdraw_queue.push(m1.clone());
-        c.withdraw_queue.push(m2.clone());
-        assert_eq!(c.resolve_withdraw_market(0).unwrap(), m1);
-        assert_eq!(c.resolve_withdraw_market(1).unwrap(), m2);
-        assert!(matches!(
-            c.resolve_withdraw_market(2),
-            Err(Error::MissingMarket(2))
-        ));
-    }
-
-    #[test]
-    fn after_supply_2_read_missing_position_stops() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        // Resolve market via supply_queue
-        let market = mk(42);
-        c.supply_queue.push(market);
-
-        // Must be in Allocating ctx
-        c.op_state = OpState::Allocating {
-            op_id: 1,
-            index: 0,
-            remaining: 10,
-        };
-
-        // Missing position -> stop_and_exit
-        let res = c.after_supply_2_read(Ok(None), 1, 0, U128(0), U128(5), U128(5));
-        match res {
-            PromiseOrValue::Value(()) => {}
-            _ => panic!("Expected Value on missing position"),
-        }
-        assert!(matches!(c.op_state, OpState::Idle));
-    }
-
-    #[test]
-    fn after_supply_2_read_read_failed_stops() {
-        let vault_id = accounts(0);
-        setup_env(&vault_id, &vault_id, vec![]);
-        let mut c = new_test_contract(&vault_id);
-
-        // Resolve market via supply_queue
-        let market = mk(43);
-        c.supply_queue.push(market);
-
-        // Must be in Allocating ctx
-        c.op_state = OpState::Allocating {
-            op_id: 7,
-            index: 0,
-            remaining: 100,
-        };
-
-        // Read failure -> stop_and_exit
-        let res = c.after_supply_2_read(
-            Err(near_sdk::PromiseError::Failed),
-            7,
-            0,
-            U128(0),
-            U128(10),
-            U128(10),
-        );
-        match res {
-            PromiseOrValue::Value(()) => {}
-            _ => panic!("Expected Value on read failure"),
-        }
-        assert!(matches!(c.op_state, OpState::Idle));
-    }
-
-    #[rstest]
-    fn after_create_withdraw_req_success_returns_promise(
-        mut c: Contract,
-        receiver: AccountId,
-        owner: AccountId,
-    ) {
-        let market = mk(50);
-        c.withdraw_queue.push(market.clone());
-        c.market_supply.insert(market.clone(), 100);
-
-        c.op_state = OpState::Withdrawing {
-            op_id: 21,
-            index: 0,
-            remaining: 60,
-            receiver: receiver.clone(),
-            collected: 10,
-            owner: owner.clone(),
-            escrow_shares: 5,
-        };
-
-        let res = c.after_create_withdraw_req(Ok(()), 21, 0, U128(60));
-        match res {
-            PromiseOrValue::Promise(_) => {}
-            _ => panic!("Expected Promise when create succeeds"),
-        }
-        // State remains Withdrawing and will continue via the promise chain
-        assert!(matches!(c.op_state, OpState::Withdrawing { .. }));
-    }
-
-    #[rstest]
-    fn after_exec_withdraw_req_returns_promise(mut c: Contract) {
-        let market = mk(60);
-        c.withdraw_queue.push(market.clone());
-        c.market_supply.insert(market.clone(), 10);
-
-        c.op_state = OpState::Withdrawing {
-            op_id: 33,
-            index: 0,
-            remaining: 5,
-            receiver: mk(9),
-            collected: 0,
-            owner: accounts(1),
-            escrow_shares: 0,
-        };
-
-        let res = c.after_exec_withdraw_req(33, 0, U128(5));
-        match res {
-            PromiseOrValue::Promise(_) => {}
-            _ => panic!("Expected Promise to read supply position after exec"),
-        }
-        assert!(matches!(c.op_state, OpState::Withdrawing { .. }));
-    }
-
-    #[rstest]
-    fn after_exec_withdraw_read_advances_when_remaining(
-        mut c: Contract,
-        owner: AccountId,
-        receiver: AccountId,
-    ) {
-        // Two markets; first has principal to withdraw
-        let m1 = mk(70);
-        let m2 = mk(71);
-        c.withdraw_queue.push(m1.clone());
-        c.withdraw_queue.push(m2.clone());
-        c.market_supply.insert(m1.clone(), 10);
-
-        c.op_state = OpState::Withdrawing {
-            op_id: 0,
-            index: 0,
-            remaining: 100,
-            receiver: receiver.clone(),
-            collected: 0,
-            owner: owner.clone(),
-            escrow_shares: 0,
-        };
-
-        // Position None => new_principal = 0 => withdrawn = 10 => credited = 10
-        let res = c.after_exec_withdraw_read(Ok(None), 0, 0, U128(10), U128(100));
-        match res {
-            PromiseOrValue::Promise(_) => {}
-            _ => panic!("Expected Promise to continue withdraw steps"),
-        }
-
-        // Idle credited, state advanced to next index with remaining reduced
-        assert_eq!(c.idle_balance, 10);
-
-        // This works
-        match &c.op_state {
-            OpState::Payout {
-                op_id,
-                receiver: r,
-                amount,
-                owner: o,
-                escrow_shares,
-                burn_shares,
-            } => {
-                assert_eq!(*op_id, 0);
-                assert_eq!(*amount, 10);
-                assert_eq!(*escrow_shares, 0);
-                assert_eq!(*burn_shares, 0);
-                assert_eq!(*r, receiver);
-                assert_eq!(*o, owner);
-            }
-            other => panic!("Unexpected state after advancing: {other:?}"),
-        }
-    }
-
-    #[rstest]
-    fn stop_and_exit_when_idle_emits_and_stays_idle(mut c: Contract) {
-        // Already Idle; ensure branch is executed
-        c.op_state = OpState::Idle;
-
-        let res = c.stop_and_exit::<&str>(Some(&"reason"));
-        match res {
-            PromiseOrValue::Value(()) => {}
-            _ => panic!("Expected Value on stop while Idle"),
-        }
-        assert!(matches!(c.op_state, OpState::Idle));
-    }
-    #[test]
-    fn accepts_increase_and_decrements_remaining() {
-        let out = reconcile_supply_outcome(&1_600, &1_000, &1_000);
-        let expected_accepted = 1_600u128.saturating_sub(1_000);
-        let expected_remaining = 1_000u128.saturating_sub(expected_accepted);
-
-        assert_eq!(out.new_principal, 1_600);
-        assert_eq!(out.accepted_event, expected_accepted); // 600
-        assert_eq!(out.remaining, expected_remaining); // 400
-    }
-
-    #[test]
-    fn no_accept_when_total_does_not_increase() {
-        // decreased
-        let out = reconcile_supply_outcome(&1_500, &2_000, &5_000);
-        assert_eq!(out.new_principal, 1_500);
-        assert_eq!(out.accepted_event, 0);
-        assert_eq!(out.remaining, 5_000);
-
-        // equal
-        let out = reconcile_supply_outcome(&2_000, &2_000, &1_234);
-        assert_eq!(out.new_principal, 2_000);
-        assert_eq!(out.accepted_event, 0);
-        assert_eq!(out.remaining, 1_234);
-    }
-
-    #[test]
-    fn remaining_saturates_to_zero_when_acceptance_exceeds_it() {
-        let out = reconcile_supply_outcome(&u128::MAX, &0, &1);
-        assert_eq!(out.new_principal, u128::MAX);
-        assert_eq!(out.accepted_event, u128::MAX);
-        assert_eq!(out.remaining, 0);
-
-        let out = reconcile_supply_outcome(&10_000, &0, &5);
-        assert_eq!(out.new_principal, 10_000);
-        assert_eq!(out.accepted_event, 10_000);
-        assert_eq!(out.remaining, 0);
-    }
-
-    #[test]
-    fn handles_extreme_boundaries_correctly() {
-        let out = reconcile_supply_outcome(&0, &0, &0);
-        assert_eq!(out.new_principal, 0);
-        assert_eq!(out.accepted_event, 0);
-        assert_eq!(out.remaining, 0);
-
-        let out = reconcile_supply_outcome(&0, &u128::MAX, &123);
-        assert_eq!(out.new_principal, 0);
-        assert_eq!(out.accepted_event, 0);
-        assert_eq!(out.remaining, 123);
-
-        let out = reconcile_supply_outcome(&u128::MAX, &(u128::MAX - 5), &2);
-        assert_eq!(out.new_principal, u128::MAX);
-        assert_eq!(out.accepted_event, 5);
-        assert_eq!(out.remaining, 0);
-    }
-
-    #[rstest]
-    fn stop_and_exit_payout_refunds_and_idle(
-        mut c: Contract,
-        owner: AccountId,
-        receiver: AccountId,
-    ) {
-        use near_sdk_contract_tools::ft::Nep141Controller as _;
-        let escrow: u128 = 10;
-
-        // Seed escrowed shares into the vault's own account
-        c.deposit_unchecked(&near_sdk::env::current_account_id(), escrow)
-            .unwrap_or_else(|e| near_sdk::env::panic_str(&e.to_string()));
-
-        // Enter Payout with non-zero escrow
-        c.op_state = OpState::Payout {
-            op_id: 123,
-            receiver: receiver.clone(),
-            amount: 77,
-            owner: owner.clone(),
-            escrow_shares: escrow,
-            burn_shares: escrow,
-        };
-
-        let supply_before = c.total_supply();
-        let vault_before = c.balance_of(&near_sdk::env::current_account_id());
-        let owner_before = c.balance_of(&owner);
-        let idle_before = c.idle_balance;
-
-        c.stop_and_exit_payout::<&str>(Some(&"reason"));
-
-        // Escrow refunded, no burn, vault goes Idle
-        assert!(matches!(c.op_state, OpState::Idle));
-        assert_eq!(c.total_supply(), supply_before, "No burn/mint on stop");
-        assert_eq!(
-            c.balance_of(&near_sdk::env::current_account_id()),
-            vault_before.saturating_sub(escrow),
-            "Vault should transfer escrow to owner"
-        );
-        assert_eq!(
-            c.balance_of(&owner),
-            owner_before.saturating_add(escrow),
-            "Owner should receive escrow refund"
-        );
-        assert_eq!(c.idle_balance, idle_before, "Idle balance unchanged");
-    }
-
-    #[rstest]
-    fn stop_and_exit_payout_zero_escrow_just_idle(
-        mut c: Contract,
-        owner: AccountId,
-        receiver: AccountId,
-    ) {
-        // Enter Payout with zero escrow; no transfers should occur
-        c.op_state = OpState::Payout {
-            op_id: 7,
-            receiver,
-            amount: 1,
-            owner: owner.clone(),
-            escrow_shares: 0,
-            burn_shares: 0,
-        };
-
-        let supply_before = c.ft_total_supply();
-        let vault_before = c.ft_balance_of(near_sdk::env::current_account_id());
-        let owner_before = c.ft_balance_of(owner.clone());
-
-        c.stop_and_exit_payout::<&str>(None);
-
-        assert!(matches!(c.op_state, OpState::Idle));
-        assert_eq!(c.ft_total_supply(), supply_before, "No supply change");
-        assert_eq!(
-            c.ft_balance_of(near_sdk::env::current_account_id()),
-            vault_before,
-            "Vault balance unchanged"
-        );
-        assert_eq!(
-            c.ft_balance_of(owner),
-            owner_before,
-            "Owner balance unchanged"
-        );
     }
 }
